Frontend - Have Charts rendering as many as needed Chart component
