+++ frontend/src/Components/Dashboards/Charts.jsx
@@ -5,12 +5,11 @@ DashboardVolunteers Component | Capstone App (Pursuit Volunteer Mgr)
 
 
 /* IMPORTS */
-
 import React, { useState, useEffect } from 'react';
-import { Bar } from 'react-chartjs-2';
+import Chart from '../Chart';
 
-export default function Chart (props) {
-    const { chartData , hoursText, eventsText } = props;
+export default function Charts (props) {
+    const { chartData } = props;
 
     const [chartInterval, setChartInterval] = useState([]);
     const [chartVolunteerHours, setChartVolunteerHours] = useState([]);
@@ -66,111 +65,113 @@ export default function Chart (props) {
     }, [chartData]);
 
 
-    // Chart for volunteer hours
-    const volunteerData = {
-        labels: chartInterval,
-        datasets: [
-            {
-                label: 'Number of Hours',
-                data: chartVolunteerHours,
-                backgroundColor: Array(12).fill('rgba(255, 99, 132, 1)'),
-                borderWidth: 2
-            }
-        ]
-    };
+    // // Chart for volunteer hours
+    // const volunteerData = {
+    //     labels: chartInterval,
+    //     datasets: [
+    //         {
+    //             label: 'Number of Hours',
+    //             data: chartVolunteerHours,
+    //             backgroundColor: Array(12).fill('rgba(255, 99, 132, 1)'),
+    //             borderWidth: 2
+    //         }
+    //     ]
+    // };
     
-    const hourOptions = {
-        options: {
-            scales: {
-                yAxes: [
-                    {
-                        ticks: {
-                            fontColor: 'white',
-                            beginAtZero: true
-                        }
-                    }
-                ],
-                xAxes: [{
-                    ticks: {
-                        fontColor: 'white',
-                        beginAtZero: true
-                    }
-                }]
-            },
-            title: {
-                display: true,
-                text: hoursText,
-                fontSize: 25,
-                fontColor: 'white'
-            },
-            legend: {
-                display: false,
-                position: 'top',
-                labels: {
-                    fontColor: 'white'
-                }
+    // const hourOptions = {
+    //     options: {
+    //         scales: {
+    //             yAxes: [
+    //                 {
+    //                     ticks: {
+    //                         fontColor: 'white',
+    //                         beginAtZero: true
+    //                     }
+    //                 }
+    //             ],
+    //             xAxes: [{
+    //                 ticks: {
+    //                     fontColor: 'white',
+    //                     beginAtZero: true
+    //                 }
+    //             }]
+    //         },
+    //         title: {
+    //             display: true,
+    //             text: hoursText,
+    //             fontSize: 25,
+    //             fontColor: 'white'
+    //         },
+    //         legend: {
+    //             display: false,
+    //             position: 'top',
+    //             labels: {
+    //                 fontColor: 'white'
+    //             }
     
-            },
-            datalabels: {
-                display: true,
-                fontColor: 'white',
-            }
-        }
-    };
+    //         },
+    //         datalabels: {
+    //             display: true,
+    //             fontColor: 'white',
+    //         }
+    //     }
+    // };
 
-    const eventData = {
-        labels: chartInterval,
-        datasets: [
-            {
-                label: 'Number of Events',
-                data: chartVolunteerEvents,
-                backgroundColor: Array(12).fill('rgba(155, 49, 117, 1)'),
-                borderWidth: 2
-            }
-        ]
-    };
+    // const eventData = {
+    //     labels: chartInterval,
+    //     datasets: [
+    //         {
+    //             label: 'Number of Events',
+    //             data: chartVolunteerEvents,
+    //             backgroundColor: Array(12).fill('rgba(155, 49, 117, 1)'),
+    //             borderWidth: 2
+    //         }
+    //     ]
+    // };
 
-    const eventOptions = {
-        options: {
-            scales: {
-                yAxes: [{
-                    ticks: {
-                        fontColor: 'white',
-                        beginAtZero: true
-                    }
-                }],
-                xAxes: [{
-                    ticks: {
-                        fontColor: 'white',
-                        beginAtZero: true
-                    }
-                }]
-            },
-            title: {
-                display: true,
-                text: eventsText,
-                fontSize: 25,
-                fontColor: 'white'
-            },
-            legend: {
-                display: false,
-                position: 'top',
-                labels: {
-                    fontColor: 'white'
-                }
-            },
-            datalabels: {
-                display: true,
-                fontColor: 'white',
-            }
-        }
-    };
+    // const eventOptions = {
+    //     options: {
+    //         scales: {
+    //             yAxes: [{
+    //                 ticks: {
+    //                     fontColor: 'white',
+    //                     beginAtZero: true
+    //                 }
+    //             }],
+    //             xAxes: [{
+    //                 ticks: {
+    //                     fontColor: 'white',
+    //                     beginAtZero: true
+    //                 }
+    //             }]
+    //         },
+    //         title: {
+    //             display: true,
+    //             text: eventsText,
+    //             fontSize: 25,
+    //             fontColor: 'white'
+    //         },
+    //         legend: {
+    //             display: false,
+    //             position: 'top',
+    //             labels: {
+    //                 fontColor: 'white'
+    //             }
+    //         },
+    //         datalabels: {
+    //             display: true,
+    //             fontColor: 'white',
+    //         }
+    //     }
+    // };
 
 
     return (
         <>
-            <Bar data={volunteerData} options={hourOptions.options} />
-            <Bar data={eventData} options={eventOptions.options} />
+            <Chart xAxes={chartInterval} data={chartVolunteerHours} 
+                title={"Events I've Participated in for a Year"} xText={"Hours"} color={'rgba(255, 99, 132, 1)'}/>
+            <Chart xAxes={chartInterval} data={chartVolunteerEvents} 
+                title={"Events I've Participated in for a Year"} xText={"Events"} color={'rgba(155, 49, 117, 1)'}/>
         </>
     )
 }
